bubble-auth/pyproject.toml
[tool.poetry]
name = "bubble-auth"
version = "0.1.0"
description = ""
authors = ["thnkr-one <jacob@thnk.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.4.0"
psycopg2-binary = "^2.9.9"
pyjwt = "^2.9.0"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
gunicorn = "^23.0.0"
google-cloud-storage = "^2.18.2"
supabase = "^2.8.1"
django-environ = "^0.11.2"
django-oauth-toolkit = "^3.0.1"
django-sslserver = "^0.22"
django-ratelimit = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
